Move gatsby functions to main process folder
